ui/requirements.txt
asn1crypto==0.23.0
backports.functools-lru-cache==1.4
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
cryptography==2.1.4
cycler==0.10.0
Django==1.11.11
django-filter==1.1.0
enum34==1.1.6
idna==2.5
inflection==0.3.1
ipaddress==1.0.18
matplotlib==2.1.1
more-itertools==4.0.1
ndg-httpsclient==0.4.3
numpy==1.13.3
pandas==0.21.0
psycopg2==2.7.4
pyasn1==0.4.2
pycparser==2.18
pyOpenSSL==17.5.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
Quandl==3.2.1
requests==2.17.3
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
SQLAlchemy==1.1.15
subprocess32==3.2.7
urllib3==1.21.1
